Created PWM folder and separated devices by communication method and Fixed PIR motion sensor delay.
