+++ src/declarative/components/PWM/VibrationMotorModule.ts
